requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
python-dotenv==1.2.1
openai==1.61.1
xdk==0.5.0
playwright==1.54.0
Pillow==11.3.0
pywin32==306
requests
requests_oauthlib